+++ src/common/mesh_processor.cpp
